✨ Use Open3 instead of system in order to capture stdout and stderr
The output of gitgud was a bit messy since git was outputting to stdout
and stderr. This was fixed by using Open3. We now capture both stdout
and stderr and check if the error actually is "remote: Internal Server
Error Everything up-to-date" before retrying. If the command was not
successful and the error isn't a match, we show the error to the user.
Also, we now have a limit of 20 retries per gitgud call.